Player bewegen und MapVerbesserungen, sind jetzt zusammengefügt.

+++ game/Game.java
@@ -16,7 +16,7 @@ public class Game extends JFrame {
 
 		display = new Display(this);
 		map = new GameMap(gameScreen.getScreenTileSize());
-		player = new Player(this, 23.5, 15, ((double)map.getTileSize() / 2), 0.07);
+		player = new Player(this, 23.5, 15, (map.getTileSize() / 2), 0.07);
 		GameKeyListener gameKeyListener = new GameKeyListener(this);
 		addKeyListener(player);
 		addKeyListener(gameKeyListener);
